+++ example/lib/screens/home_screen.dart
@@ -7,29 +7,38 @@ class HomeScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Padding(
-          padding: EdgeInsets.all(20.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Text(
-                'Welcome! Click anywhere on the screen or drag your finger '
-                'across the screen to stay logged in.',
-                textAlign: TextAlign.center,
-                style: TextStyle(fontSize: 16),
+    return Scaffold(
+      backgroundColor: Colors.white,
+      body: Stack(
+        children: [
+          Positioned(
+            right: 0,
+            left: 0,
+            top: 100,
+            child: Padding(
+              padding: const EdgeInsets.all(20.0),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.symmetric(vertical: 32.0, horizontal: 16.0),
+                    child: FlutterLogo(size: MediaQuery.of(context).size.width / 3),
+                  ),
+                  const Text(
+                    'Welcome! Click anywhere on the screen or drag your finger '
+                    'across the screen to stay logged in.',
+                    textAlign: TextAlign.center,
+                    style: TextStyle(fontSize: 16),
+                  ),
+                  const SizedBox(height: 24),
+                  const TextField(
+                    decoration: InputDecoration(labelText: 'Enter something', border: OutlineInputBorder()),
+                  ),
+                ],
               ),
-              SizedBox(height: 24),
-              TextField(
-                decoration: InputDecoration(
-                  labelText: 'Enter something',
-                  border: OutlineInputBorder(),
-                ),
-              ),
-            ],
+            ),
           ),
-        ),
+        ],
       ),
     );
   }